pyproject.toml
[project]
name = "fastapi-wasmer-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "hypercorn>=0.17.3",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "uvicorn>=0.35.0",
]
